+++ sources/feature-graph/src/main/java/ru/mertsalovda/feature_graph/view/GraphViewModel.kt
@@ -1,8 +1,6 @@
 package ru.mertsalovda.feature_graph.view
 
 import android.graphics.Color
-import android.graphics.PointF
-import android.util.Log
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -12,8 +10,8 @@ import ru.mertsalovda.core_api.interfaces.Calculator
 import ru.mertsalovda.feature_graph.di.GraphComponent
 import ru.mertsalovda.feature_graph.navigaton.ViewRouter
 import ru.mertsalovda.feature_graph.view.graph.Graph
+import ru.mertsalovda.feature_graph.view.graph.GraphUtil
 import javax.inject.Inject
-import kotlin.math.round
 
 class GraphViewModel : ViewModel() {
 
@@ -69,7 +67,8 @@ class GraphViewModel : ViewModel() {
     /** Нарисовать график */
     fun drawGraph(graphItem: GraphItem) {
         if (graphItem.isVisible) {
-            val points = getPoints(-10.00f, 10.00f, 0.10f, graphItem.expression)
+            val points =
+                GraphUtil.getPoints(-10.00f, 10.00f, 0.10f, graphItem.expression, calculator)
             _graph.postValue(Graph(graphItem.expression, points, graphItem.markerColor))
         }
     }
@@ -79,64 +78,10 @@ class GraphViewModel : ViewModel() {
         val graphItems = graphItems.value as List<GraphItem>
         val graphList = mutableListOf<Graph>()
         for (graphItem in graphItems.filter { it.isVisible }) {
-            val points = getPoints(-10.00f, 10.00f, 0.10f, graphItem.expression)
+            val points =
+                GraphUtil.getPoints(-10.00f, 10.00f, 0.10f, graphItem.expression, calculator)
             graphList.add(Graph(graphItem.expression, points, graphItem.markerColor))
         }
         _graphs.postValue(graphList)
     }
-
-    /**
-     * Получить список точек для отрисовки графика
-     *
-     * @param startX - начальное значение X,
-     * @param endX - конечное значение X,
-     * @param step - шаг изменения X,
-     * @param expression - функция
-     * @return коллекцию точек [PointF]
-     */
-    private fun getPoints(
-        startX: Float,
-        endX: Float,
-        step: Float,
-        expression: String
-    ): List<PointF?> {
-        val points = mutableListOf<PointF?>()
-        var x = startX
-        while (x <= endX) {
-            val text = replaceX(expression, round(x * 100) / 100)
-            calculator.calculate(text)
-            val point: PointF? = if (calculator.getResult().isInfinityOrNaN()) {
-                null
-            } else {
-                val roundX = round(x * 100) / 100
-                val roundY = round(calculator.getResult().toFloat() * 100) / 100
-                PointF(roundX, roundY)
-            }
-            Log.d("Point", "Point $point")
-            points.add(point)
-            x += step
-        }
-        return points
-    }
-
-    /**
-     * Подставляет в выражение вместо X переданое значение.
-     * @param expression - выражение с переменной X.
-     * @param x - значение, которое будет подставлено вместо X.
-     * @return - строка с подставленным значением вместо X.
-     */
-    private fun replaceX(expression: String, x: Float): String {
-        return if (x < 0) {
-            expression.toLowerCase()
-                .replace("+x", x.toString())
-                .replace("*x", "*(0$x)")
-                .replace("/x", "/(0$x)")
-                .replace("x", "0$x")
-        } else {
-            expression.replace("x", x.toString())
-        }
-    }
-
-    private fun Double.isInfinityOrNaN() =
-        this == Double.POSITIVE_INFINITY || this == Double.NEGATIVE_INFINITY || this.isNaN()
 }
